# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema API",
        "description": "API for Cinema",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Localhost API Server"
        }
    ],
    "paths": {
...
        "/api/sieges": {
            "get": {
                "tags": [
                    "Sièges"
                ],
                "summary": "Liste tous les sièges",
                "operationId": "531406aa02bcc8daa1024b0d0bf8d211",
                "responses": {
                    "200": {
                        "description": "Liste des sièges",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "salle_id": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "numero": {
                                                "type": "string",
                                                "example": "A1"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "standard"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Sièges"
                ],
                "summary": "Créer un nouveau siège",
                "operationId": "7f902b5096b6baa74609047df14e0c17",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "salle_id",
                                    "numero",
                                    "type"
                                ],
                                "properties": {
                                    "salle_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "numero": {
                                        "type": "string",
                                        "example": "A1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "standard",
                                            "couple"
                                        ],
                                        "example": "standard"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Siège créé avec succès"
                    }
                }
            }
        },
        "/api/sieges/{id}": {
            "get": {
                "tags": [
                    "Sièges"
                ],
                "summary": "Afficher un siège spécifique",
                "operationId": "27a0258ff8f65e9e50724cb162a6cb76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du siège",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du siège",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "salle_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "numero": {
                                            "type": "string",
                                            "example": "A1"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "standard"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Sièges"
                ],
                "summary": "Modifier un siège",
                "operationId": "ea1af9acbcca1bde2f35db805c70e2ae",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du siège",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "numero": {
                                        "type": "string",
                                        "example": "B1"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "standard",
                                            "couple"
                                        ],
                                        "example": "couple"
                                    },
                                    "reserve": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Siège mis à jour avec succès"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sièges"
                ],
                "summary": "Supprimer un siège",
                "operationId": "6315936cee7cc876c8f4186c08fae281",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du siège",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Siège supprimé avec succès"
                    }
                }
            }
        },
        "/tickets/{reservationId}": {
            "get": {
                "tags": [
                    "Tickets"
                ],
                "summary": "Generate a PDF ticket for a given reservation",
                "description": "Generates a ticket in PDF format with reservation details and QR code.",
                "operationId": "generateTicket",
                "parameters": [
                    {
                        "name": "reservationId",
                        "in": "path",
                        "description": "ID of the reservation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ticket PDF generated successfully",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "404": {
                        "description": "Reservation not found"
                    }
                }
            }
        }
    },
...
}